Implementando Login

+++ auth-service/src/controllers/UserController.ts
@@ -14,8 +14,15 @@ export default class UserController {
         
         try {
             const user: IUser = req.body;
-            const newUser = await this.userService.createUser(user);
-            res.status(201).json(newUser);
+            
+            if (!user.email && !user.password) {
+                res.status(400).send("Email e senha são obrigatórios para operação de criação de usuário");
+                return;
+            }
+
+            const newUserToken = await this.userService.createUser(user);
+            res.status(201).json(newUserToken);
+            return;
         } catch (error: any) {
             res.status(500).send(`Erro ao criar usuário: ${error}`);
         }
@@ -35,22 +42,21 @@ export default class UserController {
     }
 
 
-    public async getUserByEmail(req: Request, res: Response): Promise<void> {
-        
-        try {
-            const email = req.params.email;
-            const user = await this.userService.getUserByEmail(email);
-            res.status(200).json(user);
-        } catch (error: any) {
-            res.status(500).send(`Erro ao buscar usuário por email: ${error}`);
-        }
-
-    }
-
     public async updateUser(req: Request, res: Response): Promise<void> {
         
         try {
             const user: IUser = req.body;
+
+            if (!user.id) {
+                res.status(400).send("ID é obrigatório");
+                return;
+            }
+
+            if (!user.email && !user.password) {
+                res.status(400).send("Email ou senha são obrigatórios para operação de deleção de usuário");
+                return;
+            }
+
             const updatedUser = await this.userService.updateUser(user);
             res.status(200).json(updatedUser);
         } catch (error: any) {
@@ -63,23 +69,51 @@ export default class UserController {
     public async deleteUser(req: Request, res: Response): Promise<void> {
         
         try {
-            const id = parseInt(req.params.id);
-            await this.userService.deleteUser(id);
+
+            const user: IUser = req.body;
+
+            if (!user.id) {
+                res.status(400).send("ID é obrigatório");
+                return;
+            }
+
+            const id = parseInt(user.id);
+            const userDeleted = await this.userService.deleteUser(id);
+            
+            if(!userDeleted) {
+                res.status(401).send("Usuário não encontrado");
+                return;
+            }
+
             res.status(200).send("Usuário deletado com sucesso");
+            return;
         } catch (error: any) {
             res.status(500).send(`Erro ao deletar usuário: ${error}`);
         }
 
     }
 
-    public async getAllUsers(req: Request, res: Response): Promise<void> {
-        
+    public async login(req: Request, res: Response): Promise<void> {
+
         try {
-            const users = await this.userService.getAllUsers();
-            res.status(200).json(users);
+            const user: IUser = req.body;
+            
+            if (!user.email || !user.password) {
+                res.status(400).send("Email e senha são obrigatórios");
+                return;
+            }
+
+            const login = await this.userService.login(user);
+            
+            if(login === null) {
+                res.status(401).send("Email ou senha inválidos");
+                return;
+            }
+
+            res.status(200).json({ login });
+            return;
         } catch (error: any) {
             res.status(500).send(`Erro ao buscar todos os usuários: ${error}`);
         }
-
     }
 }
